Migrate homepage to app router
