+++ components/index.ts
@@ -1,3 +1,4 @@
 export * from './Header'
 export * from './Footer'
 export * from './Logo'
+export * from './Stars'
