+++ components/layout/layout.tsx
@@ -1,9 +1,10 @@
 import { Header } from '../Header'
 import { Footer } from '../Footer/'
+import clsx from 'clsx'
 
 const Layout: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   return (
-    <div className="mx-auto min-h-screen overflow-hidden bg-bg">
+    <div>
       <Header />
       {children}
       <Footer />
