+++ lib/get_links.tsx
@@ -1,11 +1,11 @@
 import fs from 'fs'
-import path from "path"
+import path from 'path'
 import yaml from 'js-yaml'
 
 export type LinkMetadata = {
-  name: string,
-  url: string,
-  summary: string,
+  name: string
+  url: string
+  summary: string
   subpages: undefined | LinkMetadata[]
 }
 
@@ -13,5 +13,7 @@ export function get_links() {
   const yamlLocation = path.join(process.cwd(), 'posts', 'links', 'links.yml')
   // Import
   const fileContents = fs.readFileSync(yamlLocation, 'utf-8')
-  return(yaml.load(fileContents))
+  // Dangerously coerce to type.
+  // TODO: Validate this.
+  return yaml.load(fileContents) as LinkMetadata[]
 }
