+++ next.config.js
@@ -1,16 +1,15 @@
+const withMDX = require('@next/mdx')()
+ 
 /** @type {import('next').NextConfig} */
-/*
-const withMDX = require('@next/mdx')({
-  extension: /\.mdx?$/,
-  options: {
-    remarkPlugins: [],
-    rehypePlugins: [],
-    // If you use `MDXProvider`, uncomment the following line.
-    providerImportSource: "@mdx-js/react",
-  },
-})
-*/
+const nextConfig = {
+  // Configure `pageExtensions` to include MDX files
+  pageExtensions: ['mdx', 'ts', 'tsx'],
+  // Optionally, add any other Next.js config below
+}
+ 
+module.exports = withMDX(nextConfig)
 
+/*
 module.exports = {
   webpack(config) {
     config.module.rules.push({
@@ -29,3 +28,4 @@ module.exports = {
   pageExtensions: ['ts', 'tsx', 'js', 'jsx'],
   distDir: '.next',
 }
+*/
