+++ package.json
@@ -11,11 +11,14 @@
     "start": "next start"
   },
   "dependencies": {
-    "@carbon/icons-react": "^11.32.0",
+    "@carbon/icons-react": "^11.26.0",
+    "@mdx-js/loader": "^3.0.0",
     "@mdx-js/react": "^2.3.0",
+    "@next/mdx": "^14.0.4",
+    "@types/mdx": "^2.0.10",
     "clsx": "^1.1.1",
     "eslint-config-next": "^14.0.4",
-    "framer-motion": "^10.9.1",
+    "framer-motion": "^10.16.16",
     "gray-matter": "^4.0.3",
     "js-yaml": "^4.1.0",
     "next": "^14.0.4",
@@ -26,6 +29,7 @@
   },
   "devDependencies": {
     "@svgr/webpack": "^7.0.0",
+    "@types/carbon__icons-react": "^11.26.0",
     "@types/node": "^18.15.10",
     "@types/react": "^18.0.29",
     "autoprefixer": "^10.4.0",
