--- posts/about/me.mdx
@@ -1,36 +1,0 @@
-# About me
-
-**I'm working as a UX developer for Blue Cross of Illinois**, building their Fiber design system. I'm most passionate about front-end development, though I dabble in UI and print design. Like every other software engineer on Earth, I'm also studying deep learning. For more on my skills and experience, check out my [LinkedIn](https://www.linkedin.com/in/jacob-keisling-b474a7174/).
-
-## Hobbies
-
-In my free time,
-
-- I serve as design editor for [_Animus Classics Journal_](https://voices.uchicago.edu/animus/) (check out our [fall issue](https://www.yumpu.com/en/document/view/66328266/animus-classics-journal-vol-2-issue-1)!)
-- I'm studying Japanese (currently languishing in the [intermediate plateau](https://www.reddit.com/r/languagelearning/comments/gcnryw/the_intermediate_plateau_revisited/) around JLPT N3) and know enough Latin to get by with the Vulgate or post-medieval texts.
-- I build custom keyboards - full list forthcoming
-
-## Academic interests
-
-- **ML interpretability**, both because AI notkilleveryoneism is somewhat important I guess, and because the inner representations of the world in LLMs and image models may bring insights into fundamental philosophical questions. I'm still far from actually doing meaningful interpretability research, though, and as of Q2 2023 am desperately catching up in capabilities.
-- **Technology for handling complexity:**
-  - Usability
-  - Tools for thought
-  - Information architecture
-  - Management technology
-- **History of East Asia, 1853-present,** particularly Yushin-era South Korea and occupied Manchukuo
-- **History of innovation,** particularly during the US postwar era
-
-See [questions](/writing/questions) for more on this
-
-# Colophon
-
-This website is set in [Basier Circle](https://www.atipofoundry.com/fonts/basier) and [Basier Mono](https://www.atipofoundry.com/fonts/basier-mono) typefaces. It's built with [Next.js](https://nextjs.org/) and [Tailwind](https://tailwindcss.com/) and hosted on [Cloudflare Pages](https://pages.cloudflare.com/). You can view the source code [on my GitHub](https://github.com/EndlessReform/keisling_v4) if you are so inclined.
-
-# Contact me
-
-Please don't hesitate to drop me a line! I'm happy to answer questions or meet up, either on Zoom or in Chicago. You can reach me at [contact@keisling.me](mailto:contact@keisling.me).
-
-# Disclaimer
-
-All opinions on this site are my own and do not express the views of my employer. Links do not imply endorsement unless explicitly stated.
